Added compatibility with 1.9-1.14.4.
This added compatibility with 1.9-1.14.4 all in 1 jar without needing to use a legacy jar.

I tested it with MC 1.9.4-1.14.4 and all worked. 1.8 has errors as I don't think action bar cade you use works with 1.8 but 1.9+ has no issues. Ill let you add the support for 1.8 if you wish to support it.